blog/requirements.txt
Django==5.1
django-taggit==3.0.0
psycopg2-binary==2.9.3
markdown==3.3.4
